feat: Share/Downloadのイベントはんどら
