+++ src/features/ProfileSheet/Preview/usePreview.ts
@@ -1,9 +1,7 @@
-import { RefObject, useCallback, useEffect, useState } from 'react';
+import { RefObject, useEffect } from 'react';
 
 import { toPng } from 'html-to-image';
 
-import { useSurvey } from '@/features/Survey';
-import { useDataLayer } from '@/hooks';
 import {
   useNameStore,
   useOverviewStore,
@@ -11,61 +9,7 @@ import {
   useTagStore,
   usePersonalPerspectiveStore,
 } from '@/store';
-
-type ImageOptions = {
-  share: {
-    title: string;
-    text: string;
-  };
-  fileName: string;
-};
-
-const FILENAME_USER_NAME_MAX_LENGTH = 10;
-
-// eslint-disable-next-line consistent-return
-async function base64toFile(base64url: string, options: ImageOptions) {
-  try {
-    const blob = await fetch(base64url).then((res) => res.blob());
-    return new File([blob], options.fileName, { type: blob.type });
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.log(error);
-  }
-}
-
-function saveImage(dataUrl: string, options: ImageOptions) {
-  const link = document.createElement('a');
-  link.download = options.fileName;
-  link.href = dataUrl;
-  link.click();
-}
-
-function composeUserName(userName: string) {
-  if (userName.length <= FILENAME_USER_NAME_MAX_LENGTH) {
-    return userName;
-  }
-
-  return `${userName.substring(0, FILENAME_USER_NAME_MAX_LENGTH - 1)}…`;
-}
-
-function composeImageOptions(name: string) {
-  const userName = composeUserName(name);
-
-  const sharedUrl = new URL(window.location.href);
-  sharedUrl.searchParams.set('openExternalBrowser', '1');
-  const imageOptions = {
-    share: {
-      title: `${userName}さんの自己紹介シート[こういうものです]`,
-      text: `こういうものです。自己紹介シートをお送りします。よろしくお願いいたします。
-        
-#こういうものです で自己紹介シートを作成してシェアしよう！
-${sharedUrl.toString()}`,
-    },
-    fileName: `${userName}さんの自己紹介シート_こういうものです.png`,
-  };
-
-  return imageOptions;
-}
+import { usePreviewStore } from '@/store/usePreviewStore';
 
 export function usePreview(ref: RefObject<HTMLDivElement>) {
   // TODO 更新を検知するselector
@@ -74,58 +18,7 @@ export function usePreview(ref: RefObject<HTMLDivElement>) {
   const { profileImage, coverImage } = useImageStore();
   const { tags } = useTagStore();
   const { questionAnswerPairs } = usePersonalPerspectiveStore();
-  const { scrollSurveyIntoView } = useSurvey();
-
-  const [base64url, setBase64url] = useState('');
-
-  const dataLayer = useDataLayer({
-    componentName: 'Preview',
-  });
-
-  const imageOptions = composeImageOptions(preferredName || fullName);
-
-  const handleDownload = useCallback(async () => {
-    dataLayer.pushEvent('clickDownload');
-
-    saveImage(base64url, imageOptions);
-
-    /* Survey Begin */
-    setTimeout(() => {
-      scrollSurveyIntoView();
-    }, 1000);
-    /* Survey End */
-  }, [dataLayer, base64url, imageOptions, scrollSurveyIntoView]);
-
-  const handleShare = useCallback(async () => {
-    dataLayer.pushEvent('clickShare');
-
-    const file = await base64toFile(base64url, imageOptions);
-    if (file && navigator.canShare && navigator.canShare({ files: [file] })) {
-      navigator
-        .share({
-          ...imageOptions.share,
-          files: [file],
-        })
-        .catch(() => {});
-
-      dataLayer.pushEvent('canShare');
-    } else {
-      handleDownload();
-      dataLayer.pushEvent('saveImage');
-    }
-
-    /* Survey Begin */
-    setTimeout(() => {
-      scrollSurveyIntoView();
-    }, 1000);
-    /* Survey End */
-  }, [
-    dataLayer,
-    base64url,
-    imageOptions,
-    handleDownload,
-    scrollSurveyIntoView,
-  ]);
+  const { profileSheetBase64Url, setProfileSheetBase64Url } = usePreviewStore();
 
   // The image rendering issue in Safari was addressed by implementing a workaround found at:
   // https://github.com/bubkoo/html-to-image/issues/361#issuecomment-1402537176
@@ -138,7 +31,7 @@ export function usePreview(ref: RefObject<HTMLDivElement>) {
       await toPng(ref.current, { cacheBust: true });
       toPng(ref.current, { cacheBust: true })
         .then(async (dataUrl) => {
-          setBase64url(dataUrl);
+          setProfileSheetBase64Url(dataUrl);
         })
         .catch((error) => {
           // eslint-disable-next-line no-console
@@ -159,5 +52,5 @@ export function usePreview(ref: RefObject<HTMLDivElement>) {
     questionAnswerPairs,
   ]);
 
-  return { base64url, handleShare, handleDownload };
+  return { profileSheetBase64Url };
 }
